08_Data_Versioning/scripts/requirements.txt
# Data Versioning core
dvc==2.39.0

# Data processing and analysis
pandas==2.1.4

# YAML parsing
pyyaml==6.0

# Git Python interaction
gitpython==3.1.14

# Dulwich for git plumbing, updated for compatibility
dulwich>=0.20.50

# Filesystem specification with compatible version for datasets
fsspec>=2023.1.0,<2023.11.0

# YAML runtime for great-expectations compatibility
ruamel.yaml>=0.16,<0.18

# Packaging version compliant with mlflow
packaging<24

# Timezone library version compliant with mlflow
pytz<2024

# Compat libraries for older Python versions (ignored if Python >=3.6)
pathlib2==2.3.5.post1; python_version < "3.6"
dataclasses==0.6; python_version < "3.7"
